requirements.txt
absl-py==0.4.1
astor==0.7.1
backports.weakref==1.0.post1
certifi==2018.8.24
chardet==3.0.4
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
lxml==4.2.3
Markdown==2.6.11
mock==2.0.0
nltk==3.3
numpy==1.14.5
pbr==4.2.0
Pillow==5.2.0
protobuf==3.6.1
pykml==0.1.3
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
textblob==0.15.1
urllib3==1.23
Werkzeug==0.14.1
